+++ dbmodel/schemas/main/ud/updates/4/13/0/patch.sql
@@ -8,6 +8,21 @@ or (at your option) any later version.
 
 SET search_path = SCHEMA_NAME, public, pg_catalog;
 
+ALTER TABLE gully ADD COLUMN IF NOT EXISTS dataquality INTEGER;
+ALTER TABLE gully ADD COLUMN IF NOT EXISTS dataquality_obs _int4 DEFAULT ARRAY[0];
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'dataquality', 'lyt_data_1', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'integer', 'text', 'Dataquality', 'To indicate the number of closing-opening turns when operating the valve.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature ON CONFLICT DO NOTHING;
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'dataquality_obs', 'lyt_data_1', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'text', 'text', 'Dataquality_obs', 'Observations supporting the assigned utility survey quality level.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature ON CONFLICT DO NOTHING;
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'turns_count', 'lyt_data_2', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'numeric', 'text', 'Turns count', 'To indicate the number of closing-opening turns when operating the valve.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature WHERE feature_class = 'VALVE' ON CONFLICT DO NOTHING;
+
 CREATE OR REPLACE VIEW ve_arc
 AS WITH typevalue AS (
          SELECT edit_typevalue.typevalue,