+++ dbmodel/schemas/main/ws/updates/4/13/0/patch.sql
@@ -871,3 +871,15 @@ AS WITH sector_visibility_agg AS (
      JOIN cat_element ON e.elementcat_id::text = cat_element.id::text
      LEFT JOIN sector_visibility_agg sva ON sva.element_id = e.element_id
      LEFT JOIN muni_visibility_agg mva ON mva.element_id = e.element_id;
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'dataquality', 'lyt_data_1', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'integer', 'text', 'Dataquality', 'To indicate the number of closing-opening turns when operating the valve.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature ON CONFLICT DO NOTHING;
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'dataquality_obs', 'lyt_data_1', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'text', 'text', 'Dataquality_obs', 'Observations supporting the assigned utility survey quality level.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature ON CONFLICT DO NOTHING;
+
+INSERT INTO config_form_fields (formname, formtype, tabname, columnname, layoutname, layoutorder, "datatype", widgettype, "label", tooltip, placeholder, ismandatory, isparent, iseditable, isautoupdate, isfilter, dv_querytext, dv_orderby_id, dv_isnullvalue, dv_parent_id, dv_querytext_filterc, stylesheet, widgetcontrols, widgetfunction, linkedobject, hidden, web_layoutorder) 
+SELECT child_layer, 'form_feature', 'tab_data', 'turns_count', 'lyt_data_2', (SELECT max(layoutorder)+1 FROM config_form_fields WHERE formname = child_layer AND formtype='form_feature' AND tabname='tab_data' AND layoutname='lyt_data_2'), 'numeric', 'text', 'Turns count', 'To indicate the number of closing-opening turns when operating the valve.', NULL, false, false, true, false, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, false, NULL 
+FROM cat_feature WHERE feature_class = 'VALVE' ON CONFLICT DO NOTHING;