+++ dbmodel/test/ud/schema/views/test_schema_ve_node_valve.sql
@@ -171,7 +171,7 @@ SELECT col_type_is('ve_node_valve', 'flowsetting', 'numeric(12,3)', 'Column flow
 SELECT col_type_is('ve_node_valve', 'has_access', 'bool', 'Column has_access should be bool');
 SELECT col_type_is('ve_node_valve', 'dataquality', 'int4', 'Column dataquality should be int4');
 SELECT col_type_is('ve_node_valve', 'dataquality_obs', 'int4[]', 'Column dataquality_obs should be int4[]');
-SELECT col_type_is('ve_node_valve', 'turns_count', 'numeric', 'Column turns_count should be numeric');
+SELECT col_type_is('ve_node_valve', 'turns_count', 'numeric(12,4)', 'Column turns_count should be numeric(12,4)');
 
 SELECT * FROM finish();