+++ dbmodel/test/ws/schema/views/test_schema_ve_node_outfall_valve.sql
@@ -214,7 +214,7 @@ SELECT col_type_is('ve_node_outfall_valve', 'outfallvalve_param_2', 'bool', 'Col
 
 SELECT col_type_is('ve_node_outfall_valve', 'dataquality', 'int4', 'Column dataquality should be int4');
 SELECT col_type_is('ve_node_outfall_valve', 'dataquality_obs', 'int4[]', 'Column dataquality_obs should be int4[]');
-SELECT col_type_is('ve_node_outfall_valve', 'turns_count', 'numeric', 'Column turns_count should be numeric');
+SELECT col_type_is('ve_node_outfall_valve', 'turns_count', 'numeric(12,4)', 'Column turns_count should be numeric(12,4)');
 
 SELECT * FROM finish();